file_service/requirements.txt
google-cloud-logging==3.11.3
pydantic==2.10.4
fastapi==0.115.6
uvicorn==0.34.0
absl-py==2.1.0
aiohttp==3.11.11